server-side/requirements.txt
paho-mqtt
pymongo
flask
flask-socketio
gunicorn
eventlet
flask-cors
